refactor: 총합. 재고 다듬기
