+++ src/advanced/components/stock-status-display.tsx
@@ -1,9 +1,10 @@
 import { useProductList } from '../context/product';
-import { Product } from '../data/products';
+import { getProductStockStatusString } from '../utils/product';
 
 export const StockStatusDisplay = () => {
   const { productList } = useProductList();
 
+  // 재고가 5개 미만인 상품의 상태를 문자열로 변환
   const infoMsg = productList
     // 재고가 5개 미만인 상품만 필터링
     .filter((p) => p.stockQuantity < 5)
@@ -21,13 +22,4 @@ export const StockStatusDisplay = () => {
       {infoMsg}
     </div>
   );
-};
-
-export const getProductStockStatusString = (product: Product) => {
-  const hasStock = product.stockQuantity > 0;
-  const status = hasStock
-    ? `재고 부족 (${product.stockQuantity}개 남음)`
-    : '품절';
-
-  return `${product.name}: ${status}`;
 };
